+++ CHANGELOG.md
@@ -6,6 +6,19 @@ Format jest oparty na [Keep a Changelog](https://keepachangelog.com/pl/1.1.0/). 
 
 ## [Unreleased]
 
+## [0.4.1] - 2026-08-24
+
+### Dodano
+
+- Binarkę CLI do bundla aplikacji i obrazu DMG.
+- Instalację polecenia `agentbox` w `/usr/local/bin` z poziomu ustawień aplikacji.
+- Sprawdzanie i pobieranie wyłącznie dostępnych aktualizacji skilli Git przez `agentbox update --all`.
+- Komendę `agentbox refresh`, która aktualizuje skille, tworzy backup lokalny i Git oraz synchronizuje wszystkie projekty.
+
+### Naprawiono
+
+- Dokumentację CLI, która wcześniej sugerowała, że polecenie jest dostępne globalnie bez instalacji.
+
 ## [0.4.0] - 2026-08-24
 
 ### Dodano
@@ -105,7 +118,8 @@ Format jest oparty na [Keep a Changelog](https://keepachangelog.com/pl/1.1.0/). 
 - Lokalny i zdalny backup Git biblioteki bez sekretów i ścieżek projektów.
 - Obraz instalacyjny DMG podpisany ad-hoc.
 
-[Unreleased]: https://github.com/kacperpaluch/agentbox/compare/v0.4.0...HEAD
+[Unreleased]: https://github.com/kacperpaluch/agentbox/compare/v0.4.1...HEAD
+[0.4.1]: https://github.com/kacperpaluch/agentbox/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/kacperpaluch/agentbox/compare/v0.3.2...v0.4.0
 [0.3.2]: https://github.com/kacperpaluch/agentbox/compare/v0.3.1...v0.3.2
 [0.3.1]: https://github.com/kacperpaluch/agentbox/compare/v0.3.0...v0.3.1